Revert "Moved stuff around, added dev server for testing. Updated readme"
