+++ jasperdog-discord.js
@@ -1,6 +1,5 @@
 const fs = require('fs')
-const jaspertools = require('../jasper-tools')
-const jasperService = require('./jasperdog-discord-service')
+const jaspertools = require('./jasper-tools')
 
 const Discord = require("discord.js")
 const client = new Discord.Client()
@@ -12,8 +11,13 @@ module.exports = function(){
     })
     client.on("message", msg => {
         let messageContent = msg.content.toLowerCase()
-        let image = jasperService.getImage(messageContent)
-        msg.reply(image)
+        if (messageContent.startsWith("jasper")) {
+            if(messageContent.includes("carrot")){
+                msg.reply("https://i2-prod.birminghammail.co.uk/incoming/article14143650.ece/ALTERNATES/s810/Jasper-face-crop.jpg")
+            }else{
+                msg.reply(jaspertools.getRandomImageUrl(true))
+            }
+        }
     })
     client.login(discordSecret)
 }
