+++ package.json
@@ -1,10 +1,9 @@
 {
-  "name": "jasperdog",
+  "name": "random-image",
   "version": "1.1.0",
   "license": "MIT",
   "scripts": {
-    "start": " NODE_ENV=production node app.js",
-    "dev-start": "NODE_ENV=development node app.js"
+    "start": "node app.js"
   },
   "dependencies": {
     "discord.js": "^12.1.1",
