+++ slack-auth.js
